requirements.txt
# Add your requirements here
streamlit
mysql-connector-python
matplotlib
plotly
pdf2image
google-generativeai
